@@ -1,21 +1,21 @@
 name: direct_dialer
 description: Allows Flutter applications to directly dial a phone number.
-version: 2.0.0
+version: 2.0.1
 homepage: https://github.com/GroovinChip/direct_dialer
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.5.3 <4.0.0"
   flutter: ">=1.12.0"
 
 dependencies:
   flutter:
     sdk: flutter
     
-  android_intent_plus: ^4.0.2
-  connectivity_plus: ^4.0.2
-  device_info_plus: ^9.0.3
-  permission_handler: ^10.4.3
-  url_launcher: ^6.1.12
+  android_intent_plus: ^5.2.0
+  connectivity_plus: ^6.1.0
+  device_info_plus: ^11.1.1
+  permission_handler: ^11.3.1
+  url_launcher: ^6.3.1
 
 dev_dependencies:
   flutter_test:
