@@ -10,7 +10,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -28,7 +28,7 @@ class MyApp extends StatelessWidget {
 }
 
 class DirectDialerExample extends StatefulWidget {
-  const DirectDialerExample({Key? key}) : super(key: key);
+  const DirectDialerExample({super.key});
 
   @override
   State<DirectDialerExample> createState() => _DirectDialerExampleState();
