@@ -1,9 +1,11 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:developer' as developer;
 
 import 'package:android_intent_plus/android_intent.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:device_info_plus/device_info_plus.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:url_launcher/url_launcher.dart' as url_launcher;
@@ -36,56 +38,82 @@ class DirectDialer {
   }) async {
     final phonePermission = await Permission.phone.status;
 
-    if (Platform.isAndroid) {
-      final intent = AndroidIntent(
-        action: 'android.intent.action.CALL',
-        data: 'tel:$number',
-      );
-      if (phonePermission.isDenied) {
-        Permission.phone.request().then((status) {
-          if (status.isGranted) {
-            intent.launch();
-          }
-        });
-      }
-      if (phonePermission.isGranted) {
-        try {
-          await intent.launch();
-        } catch (e) {
-          print(e);
-        }
+    if (onIpad) {
+      _dialIOS(number, retryWithFaceTime, useFaceTimeAudio);
+    } else {
+      switch (defaultTargetPlatform) {
+        case TargetPlatform.android:
+          await _dialAndroid(number, phonePermission);
+        case TargetPlatform.iOS:
+          _dialIOS(number, retryWithFaceTime, useFaceTimeAudio);
+        case TargetPlatform.macOS:
+          await dialFaceTime(number, useFaceTimeAudio);
+        default:
+          throw PlatformException(
+            code: 'direct_dialer',
+            message: 'This action is not supported on $defaultTargetPlatform',
+          );
       }
     }
+  }
 
-    if (Platform.isIOS || onIpad) {
-      // Check cellular or wifi
-      final connectivityResult = await _connectivityCheck();
-      // Check if physical device - emulators can't make calls
-      if (_iosDeviceInfo.isPhysicalDevice) {
-        // If cellular, make a call as normal
+  Future<void> _dialAndroid(
+    String number,
+    PermissionStatus phonePermission,
+  ) async {
+    final intent = AndroidIntent(
+      action: 'android.intent.action.CALL',
+      data: 'tel:$number',
+    );
+    if (phonePermission.isDenied) {
+      Permission.phone.request().then((status) {
+        if (status.isGranted) {
+          intent.launch();
+        }
+      });
+    }
+    if (phonePermission.isGranted) {
+      try {
+        await intent.launch();
+      } catch (e) {
+        print(e);
+      }
+    }
+  }
+
+  void _dialIOS(
+    String number,
+    bool retryWithFaceTime,
+    bool useFaceTimeAudio,
+  ) {
+    _onConnectivityChanged.listen((connectivityResult) {
+      if (connectivityResult.contains(ConnectivityResult.mobile)) {
         try {
-          if (connectivityResult == ConnectivityResult.mobile) {
-            await _channel.invokeMethod(
-              'dial',
-              <String, Object>{'number': number},
-            );
-          }
-        } catch (e) {
+          _channel.invokeMethod(
+            'dial',
+            <String, Object>{'number': number},
+          );
+        } catch (error, stackTrace) {
           if (retryWithFaceTime) {
-            await dialFaceTime(number, useFaceTimeAudio);
+            developer.log(
+              'Error: $error, retrying with FaceTime',
+              error: error,
+              stackTrace: stackTrace,
+            );
+            dialFaceTime(number, useFaceTimeAudio);
+          } else {
+            developer.log(
+              'Error: $error',
+              error: error,
+              stackTrace: stackTrace,
+            );
           }
         }
       } else if (_iosDeviceInfo.isPhysicalDevice &&
-          connectivityResult == ConnectivityResult.wifi) {
-        await dialFaceTime(number, useFaceTimeAudio);
-      } else {
-        throw 'This action is not available on the iOS simulator.';
+          connectivityResult.contains(ConnectivityResult.wifi)) {
+        dialFaceTime(number, useFaceTimeAudio);
       }
-    }
-
-    if (Platform.isMacOS) {
-      await dialFaceTime(number, useFaceTimeAudio);
-    }
+    });
   }
 
   /// Initiates a FaceTime call on iOS and macOS
@@ -110,8 +138,8 @@ class DirectDialer {
     }
   }
 
-  Future<ConnectivityResult> _connectivityCheck() async {
-    return await (Connectivity().checkConnectivity());
+  Stream<List<ConnectivityResult>> get _onConnectivityChanged {
+    return Connectivity().onConnectivityChanged;
   }
 
   /// Checks if the current iOS device is an iPad
